Fix thread issue. No using singleton class refactored to TaskNotificationThread.java.
